+++ package.json
@@ -12,6 +12,7 @@
     "lint": "eslint --ext .js,.vue src test/unit/specs"
   },
   "dependencies": {
+    "es6-promise": "^4.1.1",
     "vue": "^2.1.0",
     "vue-router": "^2.7.0",
     "vuex": "^2.3.1"
@@ -68,6 +69,7 @@
     "vue-style-loader": "^1.0.0",
     "vue-template-compiler": "^2.1.0",
     "vue-unit": "^0.2.2",
+    "vuenit": "^0.6.0",
     "webpack": "^1.13.2",
     "webpack-dev-middleware": "^1.8.3",
     "webpack-hot-middleware": "^2.12.2",
